requirements.txt
# requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
PyPDF2==3.0.1
openai==0.27.0
python-dotenv==0.19.0
jinja2==3.0.1
aiofiles==0.8.0
starlette==0.14.2